// src/api-generated/model/createBankAccountDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaS API
 * The core API for my multi-tenant SaaS
 * OpenAPI spec version: 1.0
 */

export interface CreateBankAccountDto {
  /** The name of the bank */
  bank_name: string;
  /** The IBAN or account number */
  iban: string;
  /** The currency of the bank account */
  currency?: string;
  /** The name of the account holder */
  account_holder_name: string;
  /** Whether this is the default bank account */
  is_default?: boolean;
}