// src/api-generated/model/updateSocialLinkDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaS API
 * The core API for my multi-tenant SaaS
 * OpenAPI spec version: 1.0
 */

export interface UpdateSocialLinkDto {
  /** The social media platform (e.g., twitter, linkedin) */
  platform?: string;
  /** The URL to the social media profile */
  url?: string;
}